buildRustPackage: optionally use vendor config from cargo-vendor
This solves a problem, where the static vendor config in
pkgs/build-support/rust/default.nix would not sufficiently replace all
crates Cargo is looking for.

In the example of #44465, crates located at a git repo would be fetched
by cargo-vendor, but not replaced in the Cargo config. Cargo then tried
to access the network, resulting in build failure.

Suggested-by: @symphorien

+++ pkgs/build-support/rust/default.nix
@@ -61,14 +61,20 @@ in stdenv.mkDerivation (args // {
     ${setupVendorDir}
 
     mkdir .cargo
-    cat >.cargo/config <<-EOF
-      [source.crates-io]
-      registry = 'https://github.com/rust-lang/crates.io-index'
-      replace-with = 'vendored-sources'
 
-      [source.vendored-sources]
-      directory = '$(pwd)/$cargoDepsCopy'
+    if [ -f "$(pwd)/$cargoDepsCopy/config" ]; then
+      sed "s|directory = \".*\"|directory = \"$(pwd)/$cargoDepsCopy\"|g" \
+        "$(pwd)/$cargoDepsCopy/config" > .cargo/config
+    else
+      cat >.cargo/config <<-EOF
+        [source.crates-io]
+        registry = 'https://github.com/rust-lang/crates.io-index'
+        replace-with = 'vendored-sources'
+
+        [source.vendored-sources]
+        directory = '$(pwd)/$cargoDepsCopy'
     EOF
+    fi
 
     unset cargoDepsCopy
